% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMLfdr.R
\name{adaptive.procedure}
\alias{adaptive.procedure}
\title{Adaptive Lfdr Procedure}
\usage{
adaptive.procedure(Y, x, pi.v, a, b, alpha = 0.05, epsilon = 0)
}
\arguments{
\item{Y}{count data}

\item{x}{covariate}

\item{pi.v}{mixing proportions/prior probabilities}

\item{a}{vector of 'a' components}

\item{b}{vector of 'b' components}

\item{alpha}{alpha level}

\item{epsilon}{epsilon value for the empirical null hypothesis}
}
\description{
Implements the adaptive Lfdr procedure to test for significant features
}
\examples{
# Generate data
data <- generate.data(M = 100, x = c(0,0,0,0,1,1,1,1), a = 2, b = c(0, 0.7),
             pi_a = c(1), pi_b = c(0.8, 0.2))

# Choose starting values
sv <- starting.values(Y = data, x = c(0,0,0,0,1,1,1,1))
a <- sv$a; b <- sv$b; pi.v <- sv$pi.v

# Get MLE estimates
EMestimates <- EM.algorithm(Y = data, x = c(0,0,0,0,1,1,1,1), start.pi = pi.v,
                  start.a = a, start.b = b)
est.a <- EMestimates$as; est.b <- EMestimates$bs; est.pi <- EMestimates$pi

# Implement procedure
procdata <- adaptive.procedure(Y = data, x = c(0,0,0,0,1,1,1,1), pi.v = est.pi,
                   a = est.a, b = est.b, epsilon = 0)
head(procdata)
}
