% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMLfdr.R
\name{Continuous.plot}
\alias{Continuous.plot}
\title{Plot for data with continuous covariate}
\usage{
Continuous.plot(lfdr.data)
}
\arguments{
\item{lfdr.data}{output from adaptive.procedure() function}
}
\description{
Creates a visual plot of MLE estimates of each feature in a data with continuous covariate
}
\examples{
# Generate data
data <- generate.data(M = 100, x = c(0,0,0,0,1,1,1,1), a = 2, b = c(0, 0.7),
                       pi_a = c(1), pi_b = c(0.8, 0.2))

# Choose starting values
sv <- starting.values(Y = data, x = c(0,0,0,0,1,1,1,1))
a <- sv$a; b <- sv$b; pi.v <- sv$pi.v

# Get MLE estimates
EMestimates <- EM.algorithm(Y = data, x = c(0,0,0,0,1,1,1,1), start.pi = pi.v,
                  start.a = a, start.b = b)
est.a <- EMestimates$as; est.b <- EMestimates$bs; est.pi <- EMestimates$pi

# Implement procedure
procdata <- adaptive.procedure(Y = data, x = c(0,0,0,0,1,1,1,1), pi.v = est.pi,
               a = est.a, b = est.b)

# Generate plot
Continuous.plot(lfdr.data = procdata$result.data)
}
