% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMLfdr.R
\name{EM.algorithm}
\alias{EM.algorithm}
\title{EM algorithm}
\usage{
EM.algorithm(
  Y,
  x,
  start.pi,
  start.a,
  start.b,
  max.it = 300,
  tol = 1e-05,
  limits = c(-10, 45)
)
}
\arguments{
\item{Y}{count data}

\item{x}{covariate}

\item{start.pi}{starting values for mixing proportions}

\item{start.a}{starting values for 'a'}

\item{start.b}{starting values for 'b'}

\item{max.it}{maximum number of iterations}

\item{tol}{tolerance level}

\item{limits}{interval to locate estimates}
}
\description{
Implements the EM algorithm to get MLE estimates
}
\examples{
# Generate data
data <- generate.data(M = 100, x = c(0,0,0,0,1,1,1,1), a = 2, b = c(0, 0.7),
           pi_a = c(1), pi_b = c(0.8, 0.2))

# Choose starting values
sv <- starting.values(Y = data, x = c(0,0,0,0,1,1,1,1))
a <- sv$a; b <- sv$b; pi.v <- sv$pi.v

# Get MLE estimates
EMestimates <- EM.algorithm(Y = data, x = c(0,0,0,0,1,1,1,1), start.pi = pi.v,
                   start.a = a, start.b = b)
EMestimates
}
