% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMLfdr.R
\name{generate.data}
\alias{generate.data}
\title{Poisson mixture data generator}
\usage{
generate.data(M, x, a, b, pi_a, pi_b)
}
\arguments{
\item{M}{number of rows to generate}

\item{x}{covariate}

\item{a}{vector of 'a' components}

\item{b}{vector of 'b' components}

\item{pi_a}{prior probability of 'a'}

\item{pi_b}{prior probability of 'b'}
}
\description{
This generates a poisson mixture data
}
\examples{
data <- generate.data(M = 100, x = c(0,0,0,0,1,1,1,1), a = 2, b = c(0, 0.7),
           pi_a = c(1), pi_b = c(0.8, 0.2))
head(data)
}
