% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMLfdr.R
\name{MA.Plot}
\alias{MA.Plot}
\title{MA plot for data with binary covariate}
\usage{
MA.Plot(lfdr.data, row_name = "Features")
}
\arguments{
\item{lfdr.data}{result.data table from adaptive.procedure() function}

\item{row_name}{label of feature names}
}
\description{
Creates an MA plot for data with binary covariate
}
\examples{
# Generate data
data <- generate.data(M = 100, x = c(0,0,0,0,1,1,1,1), a = 2, b = c(0, 0.7),
                       pi_a = c(1), pi_b = c(0.8, 0.2))

# Choose starting values
sv <- starting.values(Y = data, x = c(0,0,0,0,1,1,1,1))
a <- sv$a; b <- sv$b; pi.v <- sv$pi.v

# Get MLE estimates
EMestimates <- EM.algorithm(Y = data, x = c(0,0,0,0,1,1,1,1), start.pi = pi.v,
                  start.a = a, start.b = b)
est.a <- EMestimates$as; est.b <- EMestimates$bs; est.pi <- EMestimates$pi

# Implement procedure
procdata <- adaptive.procedure(Y = data, x = c(0,0,0,0,1,1,1,1), pi.v = est.pi,
               a = est.a, b = est.b)

# Generate plot
MA.Plot(lfdr.data = procdata$result.data)
}
