% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMLfdr.R
\name{starting.values}
\alias{starting.values}
\title{Starting values for EM algorithm}
\usage{
starting.values(Y, x, I = 1, J = 2)
}
\arguments{
\item{Y}{count data}

\item{x}{covariate}

\item{I}{number of 'a' components}

\item{J}{number of 'b' components}
}
\description{
Provides starting values to implement the EM algorithm by running a k-means clustering on the GLM estimates of each feature
}
\examples{
# Generate data
data <- generate.data(M = 100, x = c(0,0,0,0,1,1,1,1), a = 2, b = c(0, 0.7),
            pi_a = c(1), pi_b = c(0.8, 0.2))

# Choose starting values
sv <- starting.values(data, x = c(0,0,0,0,1,1,1,1))
sv

}
